updated acmp 292

+++ acmp/task-0292/code.cpp
@@ -3,7 +3,7 @@
 // Time complexity: O(sqrt(n))
 // Space complexity: O(1)
 
-bool isPrime(long long n) {
+bool isPrime(int n) {
 	for (int i = 2; i <= n / i; i++) {
 		if (n % i == 0) {
 			return false;
@@ -13,8 +13,8 @@ bool isPrime(long long n) {
 }
 
 int main() {
-	long long n;
-	scanf("%lld", &n);
+	int n;
+	scanf("%d", &n);
 	while (n > 9 && !isPrime(n)) {
 		int sum = 0;
 		while (n > 0) {
@@ -24,7 +24,7 @@ int main() {
 		n = sum;
 	}
 	if (isPrime(n)) {
-		printf("%lld", n);
+		printf("%d", n);
 	} else {
 		printf("0");
 	}
